+++ ESP32/Middleware/Services/serv_stm32_packet_framing.c
@@ -368,7 +368,7 @@ stm32_framing_status_t stm32_framing_send_packet(const uint8_t *data, size_t len
 
     // Send packet
     int sent = hal_uart_write((hal_uart_port_t)STM32_UART_PORT,
-                               tx_buffer, tx_index, timeout_ms);
+                               tx_buffer, tx_index, (int)timeout_ms);
 
     os_mutex_give(state.tx_mutex);
 
@@ -396,7 +396,7 @@ int stm32_framing_send_raw(const uint8_t *data, size_t length, uint32_t timeout_
         return -1;
     }
 
-    int sent = hal_uart_write((hal_uart_port_t)STM32_UART_PORT, data, length, timeout_ms);
+    int sent = hal_uart_write((hal_uart_port_t)STM32_UART_PORT, data, length, (int)timeout_ms);
 
     os_mutex_give(state.tx_mutex);
